Add Twitter stream processor, replies to tweets and counts users
Add a perpetual task, twitter:handle_tweets, that connects to the
Twitter Streaming API and handles incoming alerts of new tweets
and direct messages.

If the tweet mentions the user account, the bot replies that it's
happy to be thought of, and increases in redis a key associated with
the sender.

If it's a direct message, the bot also replies and increments the
redis key.

An upcoming commit will add a rake task, designed to be run hourly,
that will tweet out the person who messaged the bot the most in that
hour, or will be sad if no one messaged.