Add hourly task that determines if bot has been tweeted at
This adds a rake task, designed to be run on a periodic basis,
that determines whether it's been tweeted about, and how much,
and responds appropriately.

It queries redis for keys of the pattern "sender:*" which the
streaming twitter process `incr`'s when tweets come in.

If there are no matching keys then the bot is sad and tweets
accordingly.

If there are matching keys, it expresses its happiness at being
loved and thanks whoever tweeted at/about it the most.

It then clears redis to begin again for the next period.